feat(joins) Allow nested queries step 2b: Extract an abstract query from the logical one (#1432)
We need to make the query class generic with respect to the data source type to support nested queries but still provide a flat query to query processors. In order to get there we need to actually properly split the logical and physical query (which already have different data sources since the first relies on entities and the second on relational sources).
This brings us a step closer to that by extracting the abstract part from the query class that can be inherited by logical and physical.

It also fixes a bug in the get_all_expressions and transform_expressions methods which were ignoring the prewhere. This was not a big issue since those methods were almost never used after the prewhere condition was populated, but they may have skipped the column processor in some cases where that splitter could be applied.
